requirements.txt
APScheduler==3.6.1
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
gunicorn==19.9.0
idna==2.8
Pillow==6.2.0
psycopg2-binary==2.8.3
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tzlocal==2.0.0
urllib3==1.25.6
virtualenv==16.7.5
whitenoise==4.1.4
